+++ backend/server.js
@@ -7,17 +7,14 @@ import http from 'http';
 import { Server as socketIo } from 'socket.io';
 
 import { createTables } from './src/database/initDB.js';
-
-// Llamamos a la función para crear las tablas al arrancar el servidor
-createTables();
-
 import getPool from '../backend/src/database/getPool.js';  // Asegúrate de usar la extensión .js en la importación
 import flavorsRoutes from './routes/flavors.js';
 import ordersRoutes from './routes/orders.js';
 import coalsRoutes from './routes/coals.js';
+import tablesRoutes from '../backend/routes/tables.js'; //Nueva ruta para las mesas
 
-// Cargar las variables de entorno
-dotenv.config();  
+// Llamamos a la función para crear las tablas al arrancar el servidor
+createTables();
 
 const app = express();
 const server = http.createServer(app);  // Usamos el servidor HTTP para Socket.io
@@ -36,12 +33,6 @@ const PORT = process.env.PORT || 5000;
 app.use(cors());
 app.use(express.json());
 
-// Importar rutas
-import flavorsRoutes from './routes/flavors.js';
-import ordersRoutes from './routes/orders.js';
-import coalsRoutes from './routes/coals.js';
-import tablesRoutes from '../backend/routes/tables.js'; //Nueva ruta para las mesas
-
 // Registrar rutas
 app.use('/api/flavors', flavorsRoutes);
 app.use('/api/orders', ordersRoutes);
@@ -57,26 +48,19 @@ app.get('/', (req, res) => {
 io.on('connection', (socket) => {
   console.log('Nuevo cliente conectado');
 
-  // Recibir y manejar una nueva solicitud de carbón/cachimba
   socket.on('new-coal-request', async (data) => {
     const { tableId, requestType } = data;
 
     try {
       const pool = await getPool();
-
-      // Verificar si la mesa está libre
       const [table] = await pool.query('SELECT * FROM tables WHERE tableNumber = ?', [tableId]);
 
       if (table && table.status === 'free') {
-        // Si la mesa está libre, proceder con la solicitud y actualizar el estado de la mesa
         await pool.query('UPDATE tables SET status = "occupied" WHERE tableNumber = ?', [tableId]);
-
-        // Emitir la solicitud a todos los clientes conectados (incluyendo al jefe)
         io.emit('coal_request', data);
         console.log(`Mesa ${tableId} ha pedido ${requestType}`);
       } else {
         console.log(`Mesa ${tableId} ya está ocupada`);
-        // Aquí puedes emitir un mensaje de error si lo deseas
         socket.emit('error', `Mesa ${tableId} ya está ocupada`);
       }
     } catch (error) {
@@ -84,24 +68,17 @@ io.on('connection', (socket) => {
     }
   });
 
-  // Evento para marcar una mesa como libre
   socket.on('mark-table-free', async (tableId) => {
     try {
       const pool = await getPool();
-
-      // Verificar si la mesa existe
       const [table] = await pool.query('SELECT * FROM tables WHERE tableNumber = ?', [tableId]);
 
       if (table && table.status === 'occupied') {
-        // Si la mesa está ocupada, cambiar su estado a libre
         await pool.query('UPDATE tables SET status = "free" WHERE tableNumber = ?', [tableId]);
-
-        // Emitir el evento de cambio de estado a todos los clientes conectados
         io.emit('table-status-changed', { tableId, status: 'free' });
         console.log(`Mesa ${tableId} ha sido marcada como libre`);
       } else {
         console.log(`Mesa ${tableId} ya está libre`);
-        // Aquí también puedes emitir un mensaje si la mesa ya está libre
         socket.emit('error', `Mesa ${tableId} ya está libre`);
       }
     } catch (error) {
@@ -109,7 +86,6 @@ io.on('connection', (socket) => {
     }
   });
 
-  // Desconectar el cliente
   socket.on('disconnect', () => {
     console.log('Cliente desconectado');
   });
@@ -117,7 +93,7 @@ io.on('connection', (socket) => {
 
 // Middleware de manejo de errores
 app.use((err, req, res, next) => {
-  console.error(err); // Registrar el error en la consola
+  console.error(err);
   const status = err.status || 500;
   const message = err.message || 'Error interno del servidor';
   res.status(status).json({ error: message });
